Get broadcast working in a docker container.
Switched the base image to ubuntu instead of python3-buster because I was having issues
with it. They were probably network issues on my end but this is also a smaller base
image so that is probably better anyway. The rest of the work was just errors or missing
binds here and there to get things working. Some file paths were wrong in several places
and I needed additional bind mounts for things like the right time zone, otherwise the
time annotated on the video would be in UTC.