Instead of using curl to notify the overseer server which triggers the overseer server to request a video clip from nginx, the nginx dependency has been removed in favor of directly POSTing the video clip.
This is much cleaner all around. It removes a dependency on nginx, and removes a
volume that was shared between the containers. It reduces complexity on the overseer
side because it doesn't need to request a clip, it'll already have it. And it's
going to be hopefully more reliable as it doesn't require an individual request
for every connected client (plus one for the archiver).