If the curl request fails, retry it.
If the raspberry pi has intermittent network issues because of a flaky
wifi connection this should retry it. There's otherwise a problem with
clips being lost because the connection dropped. In a one hour archive I
usually see about 56 minutes of footage. Hopefully this fixes the issues
with lost footage.