Change how clips are received from the broadcast servers.
They now will POST a video clip to the update endpoint instead of hosting it with nginx and requiring the overseer server to make a request for each connected client and the archiver. This removed the nginx dependency from the broadcast clients, allows for the single update POST request (with automatic retries for  network issues and flaky connections), and less duplicate requests from overseer (especially with a flaky connection this should be more durable).

Relay now holds onto the most recent clips for each camera in memory for the initial immediate response when a user starts to view streams.